Add rainbow sketch functionality & name anon funcs
Clicking the "rainbow" button on the page will cause "mouseover" on any
non-black square to get colored with a random RGB value. If square is
already drawn black, it must first be erased in order to get colored
randomly.

Furthermore, anonymous functions found in addEvenListener calls in
buildGrid, erase, and resetGrid have been named and moved to separate
function declarations. This allows us to remove eventListeners when we
are switching drawing functions to ensure two different drawing events
aren't being fired at the same time, with the unintended result
overriding the intended.