Fix bug when drawing rainbow right after erasing
Bug was being caused by the drawEmpty event not being removed from the
squares when "rainbow" button was clicked, resulting in squares being
first set to BLANK_BACKGROUND before set to their random color. This
caused black squares to have the unintended effect of having their color
changed to BLANK_BACKGROUND and then the random color while on the
randomizeColor event, bypassing the if statement that first checks if a
square is black and ignores changing its color if the result is true.