Add resetRainbow
Ensures that once the rainbow mouseover event has been given to a
square, it cannot be given the event again until the board has either
been cleared or resized.

For following edits, interactivity between the opacity and rainbow
functions need improvements. While going to rainbow from opacity, the
tiles should come back as rainbow, unless they've already been colored.
As of now, this doesn't work when rainbbow is turned on after opacity,
as the opacity event still fires off, listing the squares color as black
which results in the rainbow event not firing (see end of the rainbow
function).

There may be other issues to be resolved, but as of now, that is the
most pertinent one.